Rename apple2_ir.rb to ir_simulator_runner.rb
Matches the class name (IrSimulatorRunner) and follows the same naming
convention as MOS6502's ir_simulator_runner.rb.
