+++ examples/apple2/utilities/ir_simulator_runner.rb
